+++ spec/unit/puppet/provider/bmc_network/ipmitool_spec.rb
@@ -3,20 +3,25 @@ require 'spec_helper'
 provider_class = Puppet::Type.type(:bmc_network).provider(:ipmitool)
 
 describe provider_class do
+  let :params do
+    {}
+  end
+
   let(:resource) do
-    Puppet::Type.type(:bmc_network).new(name: 'test')
+    Puppet::Type::Bmc_network.new(
+      {
+        title: 'test',
+      }.merge(params),
+    )
   end
 
   let :provider do
-    output = File.join(
-      File.dirname(__FILE__), '..', '..', '..', '..', 'fixtures', 'bmc', 'dell_ipmitool_lan_print.txt'
-    )
-    provider_class
-      .stubs(:ipmitool)
-      .returns(IO.read(output))
-    provider.prefetch('test' => resource)
-    provider
+    provider_class.new(resource)
   end
 
   it { expect(provider_class.name).to eq :ipmitool }
+
+  describe 'default instance' do
+    it { expect(provider.name).to eq 'test' }
+  end
 end